+++ src/types/module-federation.ts
@@ -15,6 +15,66 @@ export interface ModuleFederationFactoryScope {
   };
 }
 
+export interface ModuleFederationManifestV2 {
+  id: string;
+  name: string;
+  metaData: {
+    name: string;
+    type: 'app';
+    buildInfo: {
+      buildVersion: string;
+      buildName: string;
+    };
+    remoteEntry: {
+      name: string;
+      path: string;
+      type: 'global' | 'module';
+    };
+    types: {
+      path: string;
+      name: string;
+      zip: string;
+      api: string;
+    };
+    globalName: string;
+    pluginVersion: string;
+    publicPath: string;
+  };
+  shared: Array<{
+    id: string;
+    name: string;
+    version: string;
+    singleton: boolean;
+    requiredVersion: string;
+    assets: {
+      js: {
+        async: Array<string>;
+        sync: Array<string>;
+      };
+      css: {
+        sync: Array<string>;
+        async: Array<string>;
+      };
+    };
+  }>;
+  remotes: Array<any>;
+  exposes: Array<{
+    id: string;
+    name: string;
+    assets: {
+      js: {
+        sync: Array<string>;
+        async: Array<string>;
+      };
+      css: {
+        sync: Array<string>;
+        async: Array<string>;
+      };
+    };
+    path: string;
+  }>;
+}
+
 export interface ModuleFederationContainer {
   init(scope: ModuleFederationFactoryScope): void;
   get(name: string): Promise<ModuleFederationFactory>;